Fixed IntPtr.Size type is int32_t instead size_t.

+++ src/IL2C.Runtime/include/System/UIntPtr.h
@@ -37,7 +37,7 @@ extern /* virtual */ bool System_UIntPtr_Equals__System_Object(uintptr_t* this__
 #define System_UIntPtr_ToUInt32(this__) ((uint32_t)*(this__))
 #define System_UIntPtr_ToUInt64(this__) ((uint64_t)*(this__))
 
-#define System_UIntPtr_get_Size() (sizeof(uintptr_t))
+#define System_UIntPtr_get_Size() ((int32_t)sizeof(uintptr_t))
 
 #define System_UIntPtr_op_Explicit__System_UInt32(value) ((uintptr_t)(value))
 #define System_UIntPtr_op_Explicit__System_UInt64(value) ((uintptr_t)(value))
